backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
prometheus-client==0.19.0
slowapi==0.1.9
sentry-sdk[fastapi]==1.39.2
supabase==2.3.0
stripe==7.9.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
pre-commit==3.6.0